+++ spopt/__init__.py
@@ -3,6 +3,4 @@ __version__ = "0.1.0"
 
 
 # import modules/functions
-from .region_k_means import RegionKMeansHeuristic
-from .maxp import MaxPHeuristic
-from .base import  w_to_g
+from .region import (MaxPHeuristic, RegionKMeansHeuristic, w_to_g)
