+++ spopt/region/__init__.py
@@ -1,0 +1,3 @@
+from .maxp import MaxPHeuristic
+from .region_k_means import RegionKMeansHeuristic
+from .base import w_to_g
