--- spopt/region/max_p.py
@@ -1,1 +1,0 @@
-from ..base import RegionMixin
